requirements.txt
colorama==0.4.6
discord.py==2.3.0
pystyle==2.9
requests==2.28.2
ruamel.base==1.0.0
